src/main/java/br/com/mili/controller/ProdutoController.java
package br.com.mili.controller;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;

import br.com.mili.model.Produto;
import br.com.mili.service.ProdutoService;

import java.util.List;

@RestController
@RequestMapping("/produtos")
public class ProdutoController {

    @Autowired
    private ProdutoService produtoService;

    @PostMapping
    public Produto adicionarProduto(@RequestBody Produto produto) {
        return produtoService.adicionarProduto(produto);
    }

    @GetMapping
    public List<Produto> listarProdutos() {
        return produtoService.listarProdutos();
    }

    @PutMapping("/{id}")
    public Produto atualizarProduto(@PathVariable Long id, @RequestBody Produto produtoAtualizado) {
        return produtoService.atualizarProduto(id, produtoAtualizado);
    }

    @DeleteMapping("/{id}")
    public void removerProduto(@PathVariable Long id) {
        produtoService.removerProduto(id);
    }

    @GetMapping("/mais-caro")
    public Produto produtoMaisCaro() {
        return produtoService.buscarProdutoMaisCaro();
    }

    @GetMapping("/media-precos")
    public double mediaPrecos() {
        return produtoService.calcularMediaPrecos();
    }
}
